common/ukama: refactor sim type and status.
