+++ systems/common/ukama/sim_type.go
@@ -58,48 +58,3 @@ func ParseSimType(value string) SimType {
 
 	return SimType(v)
 }
-
-type SimStatus uint8
-
-const (
-	SimStatusUnknown SimStatus = iota
-	SimStatusActive
-	SimStatusInactive
-	SimStatusTerminated
-)
-
-func (s *SimStatus) Scan(value interface{}) error {
-	*s = SimStatus(uint8(value.(int64)))
-	return nil
-}
-
-func (s SimStatus) Value() (driver.Value, error) {
-	return int64(s), nil
-}
-
-func (s SimStatus) String() string {
-	t := map[SimStatus]string{0: "unknown", 1: "active", 2: "inactive", 3: "terminated"}
-
-	v, ok := t[s]
-	if !ok {
-		return t[0]
-	}
-
-	return v
-}
-
-func ParseSimStatus(value string) SimStatus {
-	i, err := strconv.Atoi(value)
-	if err == nil {
-		return SimStatus(i)
-	}
-
-	t := map[string]SimStatus{"unknown": 0, "active": 1, "inactive": 2, "terminated": 3}
-
-	v, ok := t[strings.ToLower(value)]
-	if !ok {
-		return SimStatus(0)
-	}
-
-	return SimStatus(v)
-}
