+++ systems/common/ukama/sim_type_test.go
@@ -49,37 +49,3 @@ func TestSimType(t *testing.T) {
 		assert.Equal(t, uint8(simType), uint8(10))
 	})
 }
-
-func TestSimStatus(t *testing.T) {
-	t.Run("SimTypeValidString", func(tt *testing.T) {
-		simStatus := ukama.ParseSimStatus("active")
-
-		assert.NotNil(t, simStatus)
-		assert.Equal(t, simStatus.String(), "active")
-		assert.Equal(t, uint8(simStatus), uint8(1))
-	})
-
-	t.Run("SimTypeValidNumber", func(tt *testing.T) {
-		simStatus := ukama.ParseSimStatus("3")
-
-		assert.NotNil(t, simStatus)
-		assert.Equal(t, uint8(simStatus), uint8(3))
-		assert.Equal(t, simStatus.String(), "terminated")
-	})
-
-	t.Run("SimTypeNonValidString", func(tt *testing.T) {
-		simStatus := ukama.ParseSimStatus("failure")
-
-		assert.NotNil(t, simStatus)
-		assert.Equal(t, simStatus.String(), "unknown")
-		assert.Equal(t, uint8(simStatus), uint8(0))
-	})
-
-	t.Run("SimTypeNonValidNumber", func(tt *testing.T) {
-		simStatus := ukama.SimStatus(uint8(10))
-
-		assert.NotNil(t, simStatus)
-		assert.Equal(t, simStatus.String(), "unknown")
-		assert.Equal(t, uint8(simStatus), uint8(10))
-	})
-}
